assignment1_2-apps-no-serialization/requirements.txt
django
djangorestframework
psutil
django-model-utils
